Fixing interop errors.
git-svn-id: http://localhost:81/svn/dev/modules/Uncas.Core/trunk@1163 90d2b0b2-0e11-ba47-89f1-85e69862abc4

+++ src/Uncas.Core/Interop/ScreenCapture.cs
@@ -181,12 +181,12 @@ namespace Uncas.Core.Interop
             public static extern IntPtr GetWindowDC(IntPtr hWnd);
 
             [DllImport("user32.dll")]
-            public static extern IntPtr ReleaseDC(
+            public static extern int ReleaseDC(
                 IntPtr hWnd,
                 IntPtr hDC);
 
             [DllImport("user32.dll")]
-            public static extern IntPtr GetWindowRect(
+            public static extern int GetWindowRect(
                 IntPtr hWnd,
                 ref RECT rect);
         }
